Silence MSVC 2015 type truncation warning in partial_key().
1>cuckoohash_map.hh(729): warning C4244: 'return': conversion from 'const uint16_t' to 'cuckoohash_map<...>::partial_t', possible loss of data

This commit changes the type of partial_key::hash_8bit to uint8_t, which seems
more appropriate and silences the above warning.

+++ libcuckoo/cuckoohash_map.hh
@@ -724,8 +724,8 @@ private:
                                  static_cast<uint32_t>(hash_64bit >> 32));
     const uint16_t hash_16bit = (static_cast<uint16_t>(hash_32bit) ^
                                  static_cast<uint16_t>(hash_32bit >> 16));
-    const uint16_t hash_8bit = (static_cast<uint8_t>(hash_16bit) ^
-                                static_cast<uint8_t>(hash_16bit >> 8));
+    const uint8_t hash_8bit = (static_cast<uint8_t>(hash_16bit) ^
+                               static_cast<uint8_t>(hash_16bit >> 8));
     return hash_8bit;
   }
